+++ lib/ui/LayoutBox.m
@@ -192,7 +192,7 @@
 
 -(void)removeLastSubview{
 	
-	int index=[items count]-1;
+	NSInteger index=[items count]-1;
 	if(index>-1){
 		UIView *view=[items objectAtIndex:index];
 		if([self hasSubView:view])
@@ -654,7 +654,7 @@
 		{
 			// BURightAlignMode support
 			int xpos=self.width-paddingRight;
-			int startindex=[items count]-1;
+			NSInteger startindex=[items count]-1;
 			//
 			
 			for(int i=0;i<[items count];i++){
@@ -804,7 +804,7 @@
 }
 
 
--(UIView*)viewAtIndex:(int)index{
+-(UIView*)viewAtIndex:(NSInteger)index{
 	
 	if(index<[items count]){
 		return [items objectAtIndex:index];
